Remove wrapped_globs.py
[ci skip-rust-tests]
