+++ src/python/pants/source/BUILD
@@ -2,17 +2,12 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library(
-  sources=['source_root.py', 'wrapped_globs.py'],
+  sources=['source_root.py'],
 )
 
 python_library(
   name = "filespec",
   sources = ["filespec.py"],
-)
-
-python_library(
-  name = 'payload_fields',
-  sources = ['payload_fields.py'],
 )
 
 python_tests(name='tests')
