--- src/python/pants/source/README.md
@@ -1,4 +1,0 @@
-/src/python/pants/source/
-
-Code for modeling source code, in particular its filesystem representation (as opposed to
-the dependency DAG).
